Apply workaround #60741 only if k8s version is lower than 1.11
Workaround https://github.com/kubernetes/kubernetes/pull/60741 is only
required for k8s which version is lower than 1.11.
This patch avoid to apply #60741 if k8s version is greater than 1.10.

+++ entrypoint.sh
@@ -2,9 +2,15 @@
 
 set -e -o pipefail
 
-if [[ "$(kubectl get clusterrolebinding system:node -o="go-template={{.subjects}}")" == "<no value>" ]]; then
-  echo "Deleting clusterrolebinding/system:node. This is a workaround for https://github.com/kubernetes/kubernetes/pull/60741."
-  kubectl delete clusterrolebinding system:node
+KUBE_VERSION="$(kubectl version -o json | jq -r '.serverVersion | .major+"."+.minor' | awk -F. '{printf "%2d%02d",$1,$2}')"
+KUBE_VERSION_1_11="111"
+
+# Apply workaround if k8s version is lower than 1.11.
+if [[ $KUBE_VERSION -lt $KUBE_VERSION_1_11 ]]; then
+  if [[ "$(kubectl get clusterrolebinding system:node -o="go-template={{.subjects}}")" == "<no value>" ]]; then
+    echo "Deleting clusterrolebinding/system:node. This is a workaround for https://github.com/kubernetes/kubernetes/pull/60741."
+    kubectl delete clusterrolebinding system:node
+  fi
 fi
 
 if [[ -n "$DEBUG" ]]; then
